Construction_pipeline/requirements.txt
pandas
psycopg2-binary
streamlit
plotly
python-dotenv
openpyxl
watchdog